Improve splitting of $pushAll to avoid 'have conflict mod' error
Under certain circumstances, Mongoid will generate invalid update
commands, causing MongoDB to raise an error:

  http://github.com/durran/mongoid/issues/issue/217/

This error is generally caused by trying to update a field in an
existing embeds_many document while also trying to push a new document
onto the same embeds_many array.  For example:

  $set of addresses.0.street
  $pushAll into addresses

This fails because one key is a prefix of another, and MongoDB can't
figure out that the updates don't conflict.

This fix is simple but overly conservative.  It splits the update into
two whenever a $set and a $pushAll share the same first component.  We
could obviously mirror MongoDB's actual rule more precisely, but this
would take quite a bit more Ruby code.

We include a high-level unit test in persistence_spec which will trigger
this bug at the MongoDB level.